feat(GOVBURN): auto-grow planner/runner windows via shared useMinBufferHeight hook
Extracts BURNACT's mount-time min-height mechanic (grow window body by max
content overflow) into src/hooks/use-min-buffer-height.ts and applies it to
GovBurnActWindow and GovBurnExecWindow, so the slot table, EXECUTE button,
and runner APPLY bar are always visible without scrolling.
